refactor: move common operations out of checkRun
