+++ src/api/github/checkRun.ts
@@ -1,9 +1,9 @@
 import { Octokit } from '@octokit/rest';
 import { Endpoints } from '@octokit/types';
-import fs from 'fs-extra';
-import git from 'isomorphic-git';
 
 import { pluralise } from '../../utils/logging';
+
+import { getHeadSha, getOwnerRepo } from './util';
 
 type Output = NonNullable<
   Endpoints['POST /repos/{owner}/{repo}/check-runs']['parameters']['output']
@@ -12,48 +12,6 @@ type Output = NonNullable<
 export type Annotation = NonNullable<Output['annotations']>[number];
 
 const GITHUB_MAX_ANNOTATIONS = 50;
-
-/**
- * Matches the owner and repository names in a GitHub repository URL.
- *
- * For example, given the following input strings:
- *
- * ```console
- * git@github.com:seek-oss/skuba.git
- * https://github.com/seek-oss/skuba.git
- * ```
- *
- * This pattern will produce the following matches:
- *
- * 1. seek-oss
- * 2. skuba
- */
-const ownerRepoRegex = /github.com(?::|\/)(.+)\/(.+).git$/;
-
-const getOwnerRepo = async (
-  dir: string,
-): Promise<{ owner: string; repo: string }> => {
-  const remotes = await git.listRemotes({ dir, fs });
-
-  for (const { url } of remotes) {
-    const match = ownerRepoRegex.exec(url);
-
-    const owner = match?.[1];
-    const repo = match?.[2];
-
-    if (owner && repo) {
-      return { owner, repo };
-    }
-  }
-
-  throw new Error('Could not find a GitHub remote');
-};
-
-const getHeadSha = async (dir: string): Promise<string> => {
-  const [commit] = await git.log({ depth: 1, dir, fs });
-
-  return commit.oid;
-};
 
 /**
  * Suffixes the title with the number of annotations added, e.g.
